+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 ---------
 
+v6.5 (??? 202?)
+  - Finetuned size without CustomTkinter.
+  - Set window size on Linux systems that do not support
+    state('zoomed') on tkinter windows.
+
 v6.4 (Dec 2025)
    - Draw canvas as last element so that UI controls are displayed as
      long as possible.